Version of the parser that prints only the partitions being used
This version of the parser prints on the STDOUT the partitions that are
being used by the tasks. The next step is to actually print those
informations on a separate file. Also, there are some warnings of the
type: "Use of uninitialized value in string ne at parser.pl line 65,
<FILE> line 80."