Started working on the partitions
I stopped trying to recreate the text SWF file and began reading the
data from the SWF file and putting them in arrays. The first step is to
have all the partition data in two arrays, one with the names and one
with the number of jobs for each partition.