Finished working on the contiguous version of the basic algorithm
The solution that I used on the contiguous version of the basic
algorithm was to look for valid blocks on the schedule where the new job
could be assigned. After that I sort those blocks by starting time and
take the first one.

The next step is to try to make backfilling work.