Finished fixing the backfilling code
I finished editing the backfilling code to fix the problem that happened
when the profile would say that a certain starting time was ok but that
would need the schedule to use different processors at different times.
The file conflict.swf was added to the repository to reflect that
example.