Finished coding the new test
This new test will run several threads, each one processing a few random
traces taken from the one used in the ARGV[0] parameter. The method
being used right now for creating the new traces is the block method.
The size of the block is determined by the $trace_size variable in the
beggining of the code.

I also cleaned up most of the SWF files that were in the git but weren't
being used.