Fixed a few bugs and remade some of the database code
Now the database code makes more sense. Some informations are saved in
the database for posterior use (traces and jobs) and some other things
are not saved in the database but in the csv file (cmax ratio). This way
the information from the csv can be recreated from the database but
there is no data duplication.