Attempt to fix the FCFSC code
The FCFSC code is was not respecting yet the submit times of the jobs. I
was trying to fix it but this attempt failed. Will try something
different (commiting only for saving this version for future reference).

I also started a script called property_test.pl that will be used to
test properties on the traces.