Working on the locality constraint
I implemented the locality constraint but without taking in
consideration contiguity, which means that right now I can enforce that
all processors selected for a job are inside a cluster altought they may
not be contiguous.

The next step is to test this constraint fort he FCFS and for the
Backfilling algorithms.

I also have to organize a little better the reduce_to codes in the
ProcessorSet class, they are a little messy and using several
subroutines.