I implemented a new version of the reduce_to_cluster routine
This version is meant to enforce the locality when reducing the list of
processors for a job. This routine calculates the number of clusters by
the size of the job and makes sure the job will not use more then that
number of clusters. If it is not possible, then the starting time in
question is refused.