First proposal of the new code that uses barriers
This code tries to find the last execution profile element that is
suitable for the job. That last profile can be the last one before a
barrier or the last profile, which doesn't have duration. What happens
is that jobs don't get scheduled after this barrier, which actually
limits even the execution profile itself. Since no jobs are scheduled
after this barrier, the execution profile itself ends as well. In this
case, the execution profile looks like:

- many execution profile elements with some processors available
- a gap
- the last execution profile

What happens next is that no jobs are scheduled after the gap until the
job that created the gap ends. When this happens, the gap doesn't exist
anymore (goes away from the execution profile) and new jobs are
schedule.

The only problem right now is that the code is not fair. Jobs that are
not scheduled because of a gap are put at the end of the queue. They
should go to the beginning of the queue, or something similar to that.