Testing the old version of the ProcessorRange class
This is a temporary code that I am using to test the old ProcessorRange
class. On the next commit I will remove this class and change the rest
of the code back to the new one. This commit is just to register that I
did the test and what code I used.