New version of the run subroutine in the Backfilling package
This version receives all events with the same timestamp and same type
in each call to the heap retrieve routine. Now those events are treated
together in each iteration of the run loop.

Also, were made some modifications and improvements in other packages
like ExecutionProfile, Heap and Event.

The test.swf trace in this commit is the one I used in the debugging.