Final version of the nodes_loop subroutine
This version is working as intented. First, it's only visiting the nodes
that are necessary to include all nodes in the range. Secondly, it's
receiving the subroutine as a parameter and calling it for each element
of the range. Finally, it's respecting the return from the called
subroutine so that the operation can be stopped if intended.