Fixed the nodes_loop routine and relates code
I created a new routine in the Nodes.pm for a special case when you want
to specify a routine to compare Profiles. The only case this routine is
currently used is in the remove_job routine. All the other cases use the
normal nodes_loop.

The use case tested was:
$ ./scripts/fred.pl ../swf/CEA-Curie-2011-2.1-cln-b1-clean2.swf 30 10

The next step is to test a little bit more if the code is working, then
do the other steps agreed last friday.