Implementation of an improvement to the execution time
We implemented some improvements that were aimed at the execution time
of the schedule. The most important one is regarding the n^2 nature of
the find_first_profile_for routine. The idea is to save time but only
looking at profiles that can really be used by the job and fast forward
when a bad profile is found.