Changed a little the way run_time and submitted_time are used
Before we had a routine called estimation_ending_time for jobs and
estimation_cmax. This depended on the current time. I removed that
routine and now we always use the submitted ending time. The only place
that should have access to the real ending time is the event queue. All
the rest of the code should consider the submitted ending time. That
avoids confusion and is less bug prone.