Implemented the fix for the static job delay
This delay is used when using an external simulator. Since messages
coming from the external simulator may arrive slightly later than what
we expect, we need to add some time to the submitted job ending time.
That way, even if the job uses all it's submitted time, there is some
delay for the job to finish and the event message to arrive.