Fixed a bug related to float comparison
This bug appears because we try to make:

A + B == C ?

and it seems that the numbers are very similar but not exactly
identical. So I made a small routine in Util that compares two float
numbers using a certain precision, that is a parameter. If the
difference between the number is smaller than the precision, they can be
considered to be the same.