First version of the script used to test the external simulator
This version was meant to use threads to test the execution with
external simulator several times. Right now it is only safe to use it
with one thread per time, since batsim will always use the same socket.
Eventually when it is able to use different sockets we will be able to
test it more times simultaneously.