Updated the experiment scripts
I think all the scripts are ok to go now. I am now using new processes
to run the actual schedule to avoid memory problems with Perl. It whould
take longer to finish but at least we can do it using only one script
instead of a bash loop or something like that.