New version of the experiment script
This version of the script is slightly different. To make the graph
creation easy, each instance is printed in the output file in a
different column. But now I see a problem with this if we use many
instances (i.e. 512). On the next commit I will update it again.