Add the run_mpi script
This script is responsible for running many instances of a real MPI
application in sequence. One of the inputs is the list of permutations.
These permutations are used to put the hostfile in the same order and
launch mpirun. After that the execution time is saved in an output file.