Proposal for a modification regarding new batsim message protocol
It seems that batsim has a new protocol with different messages. For
this reason, we need to make some changes to the way messages are
received. Mostly the idea is to ignore messages using the new protocol,
since they seem to be for new messages that we don't yet use.