Brought back the Platform package from a previous commit
I wanted to return to this package to try a new way to implement a
Backfilling variant that uses the platform to choose amongst the
available CPUs the best ones regarding the platform itself. This is the
version that I brought back from the previous commit, with very small
changes. From now I will make more changes and implement the idea that
I'm studying.

Amend: also brought back the Tree package.