Standardize the documentation
Before when I documented the important packages I started using the old
Perl standard. That is a tape-like standard that looks awful in the file
so that it can be converted later into man pages. I instead switched to
normal comments before the important subroutines. It will not be
possible anymore to generate help files but that have never been done,
so it's ok. It is now way more readable.