Work on the platform code
Finished working on the second idea for reducing the available CPUs to
the ones that will be used for the job using the shape of the platforms.
This second version minimizes the maximum distances between nodes used
for the job.

Another idea would be to have to versions of this algorithm. One that is
best effort and that tries to achieve this but does not delay the job
when there are enough CPUs and the forced version. The forced version
would only allow the job to execute if the maximum distance between
nodes is the minimum possible.